+++ test/beacon/Voucher.test.ts
@@ -4,11 +4,19 @@
 import { SignerWithAddress } from '@nomicfoundation/hardhat-ethers/signers';
 import { loadFixture, time } from '@nomicfoundation/hardhat-toolbox/network-helpers';
 import { expect } from 'chai';
+import { ContractTransactionReceipt } from 'ethers';
 import { ethers } from 'hardhat';
 import * as commonUtils from '../../scripts/common';
 import * as voucherHubUtils from '../../scripts/voucherHubUtils';
 import * as voucherUtils from '../../scripts/voucherUtils';
-import { IexecPocoMock, IexecPocoMock__factory, Voucher, VoucherHub } from '../../typechain-types';
+import {
+    IexecPocoMock,
+    IexecPocoMock__factory,
+    UpgradeableBeacon,
+    Voucher,
+    VoucherHub,
+    Voucher__factory,
+} from '../../typechain-types';
 import { random } from '../utils/address-utils';
 import { createMockOrder } from '../utils/poco-utils';
 
@@ -22,24 +30,45 @@ const workerpool = random();
 const appPrice = 1;
 const datasetPrice = 2;
 const workerpoolPrice = 3;
+const dealPrice = BigInt(appPrice + datasetPrice + workerpoolPrice);
 const dealId = ethers.id('deal');
 const initVoucherHubBalance = 1000; // enough to create couple vouchers
 
 describe('Voucher', function () {
+    let [
+        admin,
+        assetEligibilityManager,
+        voucherManager,
+        voucherOwner1,
+        voucherOwner2,
+        requester,
+        anyone,
+    ]: SignerWithAddress[] = [];
+    let beacon: UpgradeableBeacon;
     let iexecPoco: string;
     let iexecPocoInstance: IexecPocoMock;
     let [
-        voucherHubWithVoucherManagerSigner,
-        voucherHubWithAssetEligibilityManagerSigner,
-        voucherHubWithAnyoneSigner,
+        voucherHub,
+        voucherHubAsVoucherCreationManager,
+        voucherHubAsAssetEligibilityManager,
     ]: VoucherHub[] = [];
-    let [voucherWithOwnerSigner, voucherWithAnyoneSigner]: Voucher[] = [];
-    // We define a fixture to reuse the same setup in every test.
-    // We use loadFixture to run this setup once, snapshot that state,
-    // and reset Hardhat Network to that snapshot in every test.
+    let [
+        voucher, // TODO: Remove this when onlyAuthorized is set to matchOrders
+        voucherAsOwner,
+        voucherAsAnyone,
+    ]: Voucher[] = [];
+    let voucherCreationTxReceipt: ContractTransactionReceipt;
+    let [appOrder, datasetOrder, workerpoolOrder, requestOrder]: ReturnType<
+        typeof createMockOrder
+    >[] = [];
+
+    beforeEach('Deploy', async () => {
+        await loadFixture(deployFixture);
+    });
+
     async function deployFixture() {
         // Contracts are deployed using the first signer/account by default
-        const [
+        [
             admin,
             assetEligibilityManager,
             voucherManager,
@@ -48,26 +77,50 @@ describe('Voucher', function () {
             requester,
             anyone,
         ] = await ethers.getSigners();
-        const beacon = await voucherUtils.deployBeaconAndImplementation(admin.address);
+        // Deploy PoCo mock and VoucherHub.
         iexecPocoInstance = await new IexecPocoMock__factory()
             .connect(admin)
             .deploy()
             .then((x) => x.waitForDeployment());
         iexecPoco = await iexecPocoInstance.getAddress();
-        const voucherHub = await voucherHubUtils.deployHub(
+        beacon = await voucherUtils.deployBeaconAndImplementation(admin.address);
+        voucherHub = await voucherHubUtils.deployHub(
             assetEligibilityManager.address,
             voucherManager.address,
             iexecPoco,
             await beacon.getAddress(),
         );
-        voucherHubWithVoucherManagerSigner = voucherHub.connect(voucherManager);
-        voucherHubWithAssetEligibilityManagerSigner = voucherHub.connect(assetEligibilityManager);
-        voucherHubWithAnyoneSigner = voucherHub.connect(anyone);
-        await voucherHubWithAssetEligibilityManagerSigner.createVoucherType(description, duration);
-
+        voucherHubAsVoucherCreationManager = voucherHub.connect(voucherManager);
+        voucherHubAsAssetEligibilityManager = voucherHub.connect(assetEligibilityManager);
+        // Fund VoucherHub with RLCs.
         await iexecPocoInstance
             .transfer(await voucherHub.getAddress(), initVoucherHubBalance)
             .then((tx) => tx.wait());
+        // Create one voucher.
+        await voucherHubAsAssetEligibilityManager.createVoucherType(description, duration);
+        const voucherAddress1 = await voucherHubAsVoucherCreationManager
+            .createVoucher(voucherOwner1, voucherType, voucherValue)
+            .then((tx) => tx.wait())
+            .then((tx) => (voucherCreationTxReceipt = tx!))
+            .then(() => voucherHub.getVoucher(voucherOwner1));
+        voucher = Voucher__factory.connect(voucherAddress1, voucherOwner1);
+        voucherAsOwner = voucher.connect(voucherOwner1);
+        voucherAsAnyone = voucher.connect(anyone);
+        // Create mock orders.
+        const mockOrder = createMockOrder();
+        appOrder = { ...mockOrder, app: app, appprice: appPrice };
+        datasetOrder = {
+            ...mockOrder,
+            dataset: dataset,
+            datasetprice: datasetPrice,
+        };
+        workerpoolOrder = {
+            ...mockOrder,
+            workerpool: workerpool,
+            workerpoolprice: workerpoolPrice,
+        };
+        requestOrder = { ...mockOrder, requester: requester.address };
+        // TODO remove return and update tests.
         return {
             beacon,
             voucherHub,
@@ -83,43 +136,30 @@ describe('Voucher', function () {
 
     describe('Upgrade', function () {
         it('Should upgrade all vouchers', async function () {
-            const { beacon, voucherHub, admin, voucherOwner1, voucherOwner2 } =
-                await loadFixture(deployFixture);
-            const voucherType1 = 1;
-            const duration1 = 7200;
-            const description1 = 'Long Term Duration';
-            const voucherValue1 = 200;
-            // Create type1.
-            await voucherHubWithAssetEligibilityManagerSigner
-                .createVoucherType(description1, duration1)
-                .then((tx) => tx.wait());
-            // Create voucher1.
-            const createVoucherTx1 = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            const createVoucherReceipt1 = await createVoucherTx1.wait();
+            // Voucher1
             const expectedExpirationVoucher1 = await commonUtils.getExpectedExpiration(
                 duration,
-                createVoucherReceipt1,
+                voucherCreationTxReceipt,
             );
             const voucherAddress1 = await voucherHub.getVoucher(voucherOwner1);
             const voucherAsProxy1 = await commonUtils.getVoucherAsProxy(voucherAddress1);
-            // Create voucher2.
-            const createVoucherTx2 = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner2,
-                voucherType1,
-                voucherValue1,
-            );
-            const createVoucherReceipt2 = await createVoucherTx2.wait();
+            // Create voucher2 with different type and value.
+            const anotherDuration = 7200;
+            const anotherVoucherTypeId = 1;
+            const anotherVoucherValue = 200;
+            await voucherHubAsAssetEligibilityManager
+                .createVoucherType('Another description', anotherDuration)
+                .then((tx) => tx.wait());
+            const createVoucherReceipt2 = await voucherHubAsVoucherCreationManager
+                .createVoucher(voucherOwner2, anotherVoucherTypeId, anotherVoucherValue)
+                .then((tx) => tx.wait());
             const expectedExpirationVoucher2 = await commonUtils.getExpectedExpiration(
-                duration1,
+                anotherDuration,
                 createVoucherReceipt2,
             );
             const voucherAddress2 = await voucherHub.getVoucher(voucherOwner2);
             const voucherAsProxy2 = await commonUtils.getVoucherAsProxy(voucherAddress2);
-            // Save old implementation.
+            // Save old implementation address.
             const initialImplementation = await beacon.implementation();
             // Upgrade beacon.
             const voucherImplV2Factory = await ethers.getContractFactory('VoucherV2Mock', admin);
@@ -180,49 +220,18 @@ describe('Voucher', function () {
 
     describe('Authorization', function () {
         it('Should authorize an account', async function () {
-            const { voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-            // Authorize the account
-            const authorizationTx = await voucher
-                .connect(voucherOwner1)
-                .authorizeAccount(anyone.address);
-            const authorizationReceipt = await authorizationTx.wait();
-            // Run assertions.
-            // Events.
-            expect(authorizationReceipt)
-                .to.emit(voucher, 'AccountAuthorized')
+            expect(await voucher.isAccountAuthorized(anyone.address)).to.be.false;
+            expect(await voucherAsOwner.authorizeAccount(anyone.address))
+                .to.emit(voucherAsOwner, 'AccountAuthorized')
                 .withArgs(anyone.address);
             // Check if the account is authorized
             expect(await voucher.isAccountAuthorized(anyone.address)).to.be.true;
         });
 
-        it('Should deauthorize an account', async function () {
-            const { voucherHub, voucherOwner1, anyone } = await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-            // Authorize the account
-            await voucher.connect(voucherOwner1).authorizeAccount(anyone.address);
-            // unauthorize the account
-            const unauthorizationTx = await voucher
-                .connect(voucherOwner1)
-                .unauthorizeAccount(anyone.address);
-            const unauthorizationReceipt = await unauthorizationTx.wait();
-            // Run assertions.
-            // Events.
-            expect(unauthorizationReceipt)
+        it('Should unauthorize an account', async function () {
+            await voucherAsOwner.authorizeAccount(anyone.address).then((tx) => tx.wait());
+            expect(await voucher.isAccountAuthorized(anyone.address)).to.be.true;
+            expect(await voucherAsOwner.unauthorizeAccount(anyone.address))
                 .to.emit(voucher, 'AccountUnauthorized')
                 .withArgs(anyone.address);
             // Check if the account is unauthorized
@@ -230,75 +239,32 @@ describe('Voucher', function () {
         });
 
         it('Should not authorize account if sender is not the owner', async function () {
-            const { voucherHub, voucherManager, voucherOwner1, anyone } =
-                await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-
-            // Authorize the account
             await expect(
-                voucher.connect(anyone).authorizeAccount(anyone.address),
+                voucherAsAnyone.authorizeAccount(anyone.address),
             ).to.be.revertedWithCustomError(voucher, 'OwnableUnauthorizedAccount');
         });
 
         it('Should not unauthorize account if sender is not the owner', async function () {
-            const { voucherHub, voucherManager, voucherOwner1, anyone } =
-                await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-            await voucher.connect(voucherOwner1).authorizeAccount(anyone.address);
-            const anyoneIsAuthorized = await voucher.isAccountAuthorized(anyone.address);
+            await voucherAsOwner.authorizeAccount(anyone.address).then((tx) => tx.wait());
+            expect(await voucher.isAccountAuthorized(anyone.address)).to.be.true;
             // unauthorize the account
             await expect(
-                voucher.connect(anyone).unauthorizeAccount(anyone.address),
+                voucherAsAnyone.unauthorizeAccount(anyone.address),
             ).to.be.revertedWithCustomError(voucher, 'OwnableUnauthorizedAccount');
             // Check that the state of mapping is not modified from.
-            expect(await voucher.isAccountAuthorized(anyone.address)).to.be.equal(
-                anyoneIsAuthorized,
-            ).to.be.true;
+            expect(await voucher.isAccountAuthorized(anyone.address)).to.be.true;
         });
 
         it('Should not authorize owner account', async function () {
-            const { voucherHub, voucherManager, voucherOwner1 } = await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
+            await expect(voucherAsOwner.authorizeAccount(voucherOwner1.address)).to.be.revertedWith(
+                'Voucher: owner is already authorized.',
             );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-
-            await expect(
-                voucher.connect(voucherOwner1).authorizeAccount(voucherOwner1.address),
-            ).to.be.revertedWith('Voucher: owner is already authorized.');
         });
     });
 
     describe('Voucher Balance', function () {
         it('Should get balance', async function () {
-            const { voucherHub, voucherOwner1 } = await loadFixture(deployFixture);
-            const createVoucherTx = await voucherHubWithVoucherManagerSigner.createVoucher(
-                voucherOwner1,
-                voucherType,
-                voucherValue,
-            );
-            await createVoucherTx.wait();
-            const voucherAddress = await voucherHub.getVoucher(voucherOwner1);
-            const voucher: Voucher = await commonUtils.getVoucher(voucherAddress);
-            expect(await voucher.getBalance()).to.be.equal(voucherValue);
+            expect(await voucherAsAnyone.getBalance()).to.be.equal(voucherValue);
         });
     });
 
@@ -307,42 +273,9 @@ describe('Voucher', function () {
             iexecPocoInstance.balanceOf(voucher.getAddress());
         const getRequesterBalanceOnIexecPoco = () =>
             iexecPocoInstance.balanceOf(requester.getAddress());
-        const dealPrice = BigInt(appPrice + datasetPrice + workerpoolPrice);
-        const mockOrder = createMockOrder();
-        const appOrder = { ...mockOrder, app: app, appprice: appPrice };
-        const datasetOrder = {
-            ...mockOrder,
-            dataset: dataset,
-            datasetprice: datasetPrice,
-        };
-        const workerpoolOrder = {
-            ...mockOrder,
-            workerpool: workerpool,
-            workerpoolprice: workerpoolPrice,
-        };
-        let requestOrder = { ...mockOrder };
-        let [voucherOwner1, requester, anyone]: SignerWithAddress[] = [];
-        let voucherHub: VoucherHub;
-        let voucher: Voucher; // TODO: Remove this when onlyAthorized is set to matchOrders
-
-        beforeEach(async () => {
-            ({ voucherHub, voucherOwner1, requester, anyone } = await loadFixture(deployFixture));
-            requestOrder.requester = requester.address;
-            voucher = await voucherHubWithVoucherManagerSigner
-                .createVoucher(voucherOwner1, voucherType, voucherValue)
-                .then((tx) => tx.wait())
-                .then(() => voucherHub.getVoucher(voucherOwner1))
-                .then((voucherAddress) => commonUtils.getVoucher(voucherAddress));
-            voucherWithOwnerSigner = voucher.connect(voucherOwner1);
-            voucherWithAnyoneSigner = voucher.connect(anyone);
-        });
 
         it('Should match orders with full sponsored amount', async () => {
-            for (const asset of [app, dataset, workerpool]) {
-                await voucherHubWithAssetEligibilityManagerSigner
-                    .addEligibleAsset(voucherType, asset)
-                    .then((x) => x.wait());
-            }
+            await addEligibleAssets([app, dataset, workerpool]);
             const voucherInitialCreditBalance = await voucher.getBalance();
             const voucherInitialSrlcBalance = await getVoucherBalanceOnIexecPoco();
             const requesterInitialSrlcBalance = await getRequesterBalanceOnIexecPoco();
@@ -415,16 +348,12 @@ describe('Voucher', function () {
         describe('Match orders boost', async function () {
             it('Should match orders boost with full sponsored amount', async () => {
                 const sponsoredValue = BigInt(appPrice + datasetPrice + workerpoolPrice);
-                for (const asset of [app, dataset, workerpool]) {
-                    await voucherHubWithAssetEligibilityManagerSigner
-                        .addEligibleAsset(voucherType, asset)
-                        .then((x) => x.wait());
-                }
+                await addEligibleAssets([app, dataset, workerpool]);
                 const voucherInitialCreditBalance = await voucher.getBalance();
                 const voucherInitialSrlcBalance = await getVoucherBalanceOnIexecPoco();
                 const requesterInitialSrlcBalance = await getRequesterBalanceOnIexecPoco();
                 expect(
-                    await voucherWithOwnerSigner.matchOrdersBoost.staticCall(
+                    await voucherAsOwner.matchOrdersBoost.staticCall(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -432,7 +361,7 @@ describe('Voucher', function () {
                     ),
                 ).to.be.equal(dealId);
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -469,7 +398,7 @@ describe('Voucher', function () {
                 const requesterInitialSrlcBalance = await getRequesterBalanceOnIexecPoco();
 
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -492,11 +421,7 @@ describe('Voucher', function () {
             it('Should match orders boost with partial sponsored amount', async () => {
                 const sponsoredValue = BigInt(datasetPrice + workerpoolPrice);
                 const noSponsoredValue = BigInt(appPrice); // app wont be eligible for sponsoring
-                for (const asset of [dataset, workerpool]) {
-                    await voucherHubWithAssetEligibilityManagerSigner
-                        .addEligibleAsset(voucherType, asset)
-                        .then((x) => x.wait());
-                }
+                await addEligibleAssets([dataset, workerpool]);
                 const voucherInitialCreditBalance = await voucher.getBalance();
                 const voucherInitialSrlcBalance = await getVoucherBalanceOnIexecPoco();
                 const requesterInitialSrlcBalance = await getRequesterBalanceOnIexecPoco();
@@ -514,7 +439,7 @@ describe('Voucher', function () {
                     .then((tx) => tx.wait());
 
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -544,17 +469,11 @@ describe('Voucher', function () {
             });
 
             it('Should match orders boost with an authorized account', async () => {
-                for (const asset of [app, dataset, workerpool]) {
-                    await voucherHubWithAssetEligibilityManagerSigner
-                        .addEligibleAsset(voucherType, asset)
-                        .then((x) => x.wait());
-                }
-                await voucherWithOwnerSigner
-                    .authorizeAccount(anyone.address)
-                    .then((tx) => tx.wait());
+                await addEligibleAssets([app, dataset, workerpool]);
+                await voucherAsOwner.authorizeAccount(anyone.address).then((tx) => tx.wait());
 
                 await expect(
-                    voucherWithAnyoneSigner.matchOrdersBoost(
+                    voucherAsAnyone.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -567,7 +486,7 @@ describe('Voucher', function () {
 
             it('Should not match orders boost when sender is not allowed', async () => {
                 await expect(
-                    voucherWithAnyoneSigner.matchOrdersBoost(
+                    voucherAsAnyone.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -580,7 +499,7 @@ describe('Voucher', function () {
                 const expirationDate = await voucher.getExpiration();
                 await time.setNextBlockTimestamp(expirationDate);
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -591,7 +510,7 @@ describe('Voucher', function () {
 
             it('Should not match orders boost when non-sponsored amount is not transferable', async () => {
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         appOrder,
                         datasetOrder,
                         workerpoolOrder,
@@ -607,7 +526,7 @@ describe('Voucher', function () {
                     .willRevertOnSponsorMatchOrdersBoost()
                     .then((tx) => tx.wait());
                 await expect(
-                    voucherWithOwnerSigner.matchOrdersBoost(
+                    voucherAsOwner.matchOrdersBoost(
                         { ...appOrder, appprice: 0 },
                         { ...datasetOrder, datasetprice: 0 },
                         { ...workerpoolOrder, workerpoolprice: 0 },
@@ -617,4 +536,12 @@ describe('Voucher', function () {
             });
         });
     });
+
+    async function addEligibleAssets(assets: string[]) {
+        for (const asset of assets) {
+            await voucherHubAsAssetEligibilityManager
+                .addEligibleAsset(voucherType, asset)
+                .then((tx) => tx.wait());
+        }
+    }
 });