+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Refactor voucher tests file. (#18)
 - Use real poco address if available at deployment. (#17)
 - Match orders boost through voucher. (#16)
 - Use hardhat deploy. (#15)